backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.11"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.14",
    "logfire[fastapi,sqlite3]>=3.22.0",
    "pydantic-ai[logfire]>=0.3.5",
    "python-dotenv>=1.1.1",
    "redis>=6.2.0",
    "tavily-python>=0.7.9",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "ruff>=0.12.1",
    "taskipy>=1.14.1",
]

[tool.taskipy.tasks]
dev = "uv run uvicorn app.main:app --host 0.0.0.0 --port 8000 --reload"
lint = "uv run ruff check"
start = "uv run uvicorn app.main:app --host 0.0.0.0 --port 8000"
test = "uv run pytest"
